Include filename in hash of file.
It is possible for two separate files to have the same contents but
different data about the file, like path, what module it belongs
to, etc.

This includes the filename and the file contents in the hash that
we calculate to determine uniqueness, so that different files with
the same contents will be cached in different files.